requirements.txt
playwright==1.41.0
anthropic>=0.8.1
schedule==1.2.0
python-dotenv==1.0.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.86.0
requests==2.31.0